Add support for comments
Closes #11
