+++ test/test.html
@@ -16,7 +16,7 @@
   import "./test-event.js";
   import "./test-cascade.js";
   import "./test-custom-property.js";
-  import "./test-selector.js";
+  import "./test-language.js";
   import "./test-each.js";
   import "./test-prop.js";
   import "./test-attr.js";
